backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt<5
email-validator==2.2.0
langchain>=1.2,<1.3
langchain-openai>=1.0,<1.1
python-dotenv==1.1.1
redis==5.2.1
psycopg[binary]==3.2.9
chromadb==1.0.20
